Other changes include two reverts so that the ROM can be compiled.

+++ init_los20.sh
@@ -1,15 +1,33 @@
 #!/bin/bash
 
-rm -rf .repo/manifests
+if [ -d .repo/manifests ] ; then
+  rm -rf .repo/manifests
+fi
+
+if [ -d .repo/repo ] ; then
+  rm -rf .repo/repo
+fi
 
 repo init -u https://github.com/LineageOS-UL/android.git -b lineage-20.0 --git-lfs
 
 repo sync --force-sync
 
+# *******************
+# * Error handling: *
+# *******************
+
+status=$?
+if [ $status -eq 1 ]; then
+  echo "$(tput setaf 1)############################################$(tput setaf 0)"
+  echo "$(tput setaf 1)# An error occurred during initialization! #$(tput setaf 0)"
+  echo "$(tput setaf 1)############################################$(tput setaf 0)"
+  exit
+fi
+
 if [ -d vendor/fdroid ] ; then
-	cd vendor/fdroid
-		./get_packages.sh
-	cd ../..
+  cd vendor/fdroid
+    ./get_packages.sh
+  cd ../..
 fi
 
 if [ -d vendor/microG ] ; then
@@ -19,13 +37,21 @@ if [ -d vendor/microG ] ; then
 fi
 
 cd frameworks/base
-	git fetch "https://github.com/K9100ii/android_frameworks_base" lineage-20.0 && git cherry-pick 8c6aabf38f8cc93f34638a8ebdb9bd286ba847b7
+  git fetch "https://github.com/K9100ii/android_frameworks_base" lineage-20.0 && git cherry-pick 8c6aabf38f8cc93f34638a8ebdb9bd286ba847b7
 cd ../..
 
 cd hardware/qcom-caf/msm8994/audio/
-	git fetch "https://github.com/LineageOS/android_hardware_qcom_audio" lineage-18.1-caf-msm8916 && git cherry-pick ce771ebbc1988db8ccefb20615a8ca6295c3e93a
+  git fetch "https://github.com/LineageOS/android_hardware_qcom_audio" lineage-18.1-caf-msm8916 && git cherry-pick ce771ebbc1988db8ccefb20615a8ca6295c3e93a
 cd ../../../..
 
 cd packages/modules/Permission/
-	git fetch "https://github.com/K9100ii/android_packages_modules_Permission" lineage-20.0 && git cherry-pick 35c3ad52a0b1af32d308762bd8c63f3eb5a02aba
+  git fetch "https://github.com/K9100ii/android_packages_modules_Permission" lineage-20.0 && git cherry-pick 35c3ad52a0b1af32d308762bd8c63f3eb5a02aba
 cd ../../..
+
+cd packages/providers/DownloadProvider
+  git revert a16ad15fffeb309037bd964b61b3c5e4a8251298
+cd ../../..
+
+cd vendor/qcom/opensource/vibrator/
+  git revert 61cfb7a5d55f6183d9d583bb30712e48d45b12c9
+cd ../../../..
